fix(tools): address shell_compact review feedback
